Consolidate schedule page
